+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - CI-250 - Added an opportunity listing page template and put it behind a feature flag
 - No ticket - Added 'Opportunity' Subheading for opportunity pages
 - No ticket - Update breadcrumbs on opportunity pages.
+- CI-271 - Add new URL for 'Invest Capital Homepage'
 
 ## [2019.05.16](https://github.com/uktrade/great-international-ui/releases/tag/2019.05.16)
 [Full Changelog](https://github.com/uktrade/great-international-ui/compare/2019.05.13_2...2019.05.16)